requirements.txt
fastapi==0.115.7
pydantic==2.10.6
uvicorn==0.34.0
python-multipart==0.0.20
transformers==4.51.3
accelerate==1.3.0
torch==2.7.0
ffmpeg-python==0.2.0
pillow==11.1.0
bitsandbytes==0.45.1
python-dotenv==1.0.1
qwen-vl-utils==0.0.10
torchvision==0.20.1
ollama==0.4.7
pydantic-settings==2.3.0
pytest==8.2.2
httpx==0.27.0
